+++ api/middlewares/middlewares.authorization.js
@@ -1,5 +1,5 @@
-const { decodeToken } = require('../utils.cryptography')
-const services = require('../../services/services.user')
+const { decodeToken } = require('../utils/utils.cryptography')
+const services = require('../services/services.user')
 
 const authorizationMiddleware = (permission) => async (req, res, next) => {
   const { token } = req.headers
