+++ api/routes/v1/routes.v1.category.js
@@ -1,11 +1,11 @@
 const joi = require('joi')
 const controllerCategory = require('../../controllers/controllers.category')
 
-const middlewareValidateDTO = require('../../utils/middlewares/middlewares.validate-dto')
-const middlewareFileUploadMiddleware = require('../../utils/middlewares/middlewares.file-upload')
-const authenticationMiddleware = require('../../utils/middlewares/middlewares.authentication')
-const authorizationMiddleware = require('../../utils/middlewares/middlewares.authorization')
-const verifyDbMiddleware = require('../../utils/middlewares/middlewares.verify-exists')
+const middlewareValidateDTO = require('../../middlewares/middlewares.validate-dto')
+const middlewareFileUploadMiddleware = require('../../middlewares/middlewares.file-upload')
+const authenticationMiddleware = require('../../middlewares/middlewares.authentication')
+const authorizationMiddleware = require('../../middlewares/middlewares.authorization')
+const verifyDbMiddleware = require('../../middlewares/middlewares.verify-exists')
 
 module.exports = (router) => {
   router
