+++ api/routes/v1/routes.v1.provider.js
@@ -1,9 +1,9 @@
 const joi = require('joi')
 const providerController = require('../../controllers/controllers.provider')
-const middlewareValidateDTO = require('../../utils/middlewares/middlewares.validate-dto')
-const authenticationMiddleware = require('../../utils/middlewares/middlewares.authentication')
-const authorizationMiddleware = require('../../utils/middlewares/middlewares.authorization')
-const verifyDbMiddleware = require('../../utils/middlewares/middlewares.verify-exists')
+const middlewareValidateDTO = require('../../middlewares/middlewares.validate-dto')
+const authenticationMiddleware = require('../../middlewares/middlewares.authentication')
+const authorizationMiddleware = require('../../middlewares/middlewares.authorization')
+const verifyDbMiddleware = require('../../middlewares/middlewares.verify-exists')
 
 module.exports = (router) => {
   router.route('/provider/filter/:namefilter').get(
