+++ api/routes/v1/routes.v1.user.js
@@ -1,6 +1,6 @@
 const joi = require('joi')
 const userController = require('../../controllers/controllers.user')
-const middlewareValidateDTO = require('../../utils/middlewares/middlewares.validate-dto')
+const middlewareValidateDTO = require('../../middlewares/middlewares.validate-dto')
 
 module.exports = (router) => {
   router.route('/auth').post(
